Renames `testing` folders to `__data__`
